// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  TECHNICIEN: 'TECHNICIEN',
  VIEWER: 'VIEWER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const SiteStatus = {
  ACTIF: 'ACTIF',
  EN_CONSTRUCTION: 'EN_CONSTRUCTION',
  INACTIF: 'INACTIF'
} as const

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus]


export const EquipementStatus = {
  EN_SERVICE: 'EN_SERVICE',
  EN_MAINTENANCE: 'EN_MAINTENANCE',
  HORS_SERVICE: 'HORS_SERVICE'
} as const

export type EquipementStatus = (typeof EquipementStatus)[keyof typeof EquipementStatus]


export const Priority = {
  CRITIQUE: 'CRITIQUE',
  HAUTE: 'HAUTE',
  MOYENNE: 'MOYENNE',
  FAIBLE: 'FAIBLE'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const ChantierStatus = {
  A_PLANIFIER: 'A_PLANIFIER',
  PLANIFIE: 'PLANIFIE',
  EN_COURS: 'EN_COURS',
  EN_ATTENTE: 'EN_ATTENTE',
  TERMINE: 'TERMINE'
} as const

export type ChantierStatus = (typeof ChantierStatus)[keyof typeof ChantierStatus]


export const MaintenanceType = {
  PREVENTIVE: 'PREVENTIVE',
  CORRECTIVE: 'CORRECTIVE'
} as const

export type MaintenanceType = (typeof MaintenanceType)[keyof typeof MaintenanceType]


export const MaintenanceStatus = {
  A_PLANIFIER: 'A_PLANIFIER',
  PLANIFIEE: 'PLANIFIEE',
  EN_COURS: 'EN_COURS',
  EN_ATTENTE: 'EN_ATTENTE',
  TERMINEE: 'TERMINEE'
} as const

export type MaintenanceStatus = (typeof MaintenanceStatus)[keyof typeof MaintenanceStatus]


export const TaskStatus = {
  A_FAIRE: 'A_FAIRE',
  EN_COURS: 'EN_COURS',
  EN_ATTENTE: 'EN_ATTENTE',
  TERMINEE: 'TERMINEE'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]